+++ egui/src/memory.rs
@@ -296,15 +296,21 @@ impl Memory {
         !self.had_focus_last_frame(id) && self.has_focus(id)
     }
 
-    pub(crate) fn has_focus(&self, id: Id) -> bool {
+    /// Does this widget have keybaord focus?
+    #[inline(always)]
+    pub fn has_focus(&self, id: Id) -> bool {
         self.interaction.focus.id == Some(id)
     }
 
-    /// Give keyboard focus to a specific widget
+    /// Give keyboard focus to a specific widget.
+    /// See also [`crate::Response::request_focus`].
+    #[inline(always)]
     pub fn request_focus(&mut self, id: Id) {
         self.interaction.focus.id = Some(id);
     }
 
+    /// Surrender keyboard focus for a specific widget.
+    /// See also [`crate::Response::surrender_focus`].
     #[inline(always)]
     pub fn surrender_focus(&mut self, id: Id) {
         if self.interaction.focus.id == Some(id) {
@@ -320,14 +326,17 @@ impl Memory {
     }
 
     /// Stop editing of active `TextEdit` (if any).
+    #[inline(always)]
     pub fn stop_text_input(&mut self) {
         self.interaction.focus.id = None;
     }
 
+    #[inline(always)]
     pub fn is_anything_being_dragged(&self) -> bool {
         self.interaction.drag_id.is_some()
     }
 
+    #[inline(always)]
     pub fn is_being_dragged(&self, id: Id) -> bool {
         self.interaction.drag_id == Some(id)
     }